change checking for TravelCertificate destination availability
