+++ app/data/data-travel-certificate/data-travel-certificate-public/src/main/kotlin/com/hedvig/android/data/travelcertificate/di/TravelCertificateModule.kt
@@ -3,8 +3,6 @@ package com.hedvig.android.data.travelcertificate.di
 import com.apollographql.apollo3.ApolloClient
 import com.hedvig.android.data.travelcertificate.CheckTravelCertificateAvailabilityForCurrentContractsUseCase
 import com.hedvig.android.data.travelcertificate.CheckTravelCertificateAvailabilityForCurrentContractsUseCaseImpl
-import com.hedvig.android.data.travelcertificate.CheckTravelCertificateDestinationAvailabilityUseCase
-import com.hedvig.android.data.travelcertificate.CheckTravelCertificateDestinationAvailabilityUseCaseImpl
 import com.hedvig.android.data.travelcertificate.GetCoInsuredForContractUseCase
 import com.hedvig.android.data.travelcertificate.GetCoInsuredForContractUseCaseImpl
 import com.hedvig.android.data.travelcertificate.GetEligibleContractsWithAddressUseCase
@@ -24,12 +22,6 @@ val travelCertificateDataModule = module {
     GetTravelCertificatesHistoryUseCaseImpl(
       get<ApolloClient>(),
       get<Clock>(),
-    )
-  }
-  single<CheckTravelCertificateDestinationAvailabilityUseCase> {
-    CheckTravelCertificateDestinationAvailabilityUseCaseImpl(
-      get<GetTravelCertificatesHistoryUseCase>(),
-      get<CheckTravelCertificateAvailabilityForCurrentContractsUseCase>(),
     )
   }
   single<CheckTravelCertificateAvailabilityForCurrentContractsUseCase> {
