+++ app/feature/feature-profile/src/main/graphql/member/QueryTravelCertificateAvailability.graphql
@@ -1,0 +1,7 @@
+query TravelCertificateAvailability {
+  currentMember {
+    memberActions {
+      isTravelCertificateEnabled
+    }
+  }
+}
