+++ app/feature/feature-profile/src/main/kotlin/com/hedvig/android/feature/profile/di/ProfileModule.kt
@@ -6,7 +6,8 @@ import com.hedvig.android.apollo.auth.listeners.UploadLanguagePreferenceToBacken
 import com.hedvig.android.auth.LogoutUseCase
 import com.hedvig.android.core.demomode.DemoManager
 import com.hedvig.android.data.settings.datastore.SettingsDataStore
-import com.hedvig.android.data.travelcertificate.CheckTravelCertificateDestinationAvailabilityUseCase
+import com.hedvig.android.feature.profile.data.CheckTravelCertificateDestinationAvailabilityUseCase
+import com.hedvig.android.feature.profile.data.CheckTravelCertificateDestinationAvailabilityUseCaseImpl
 import com.hedvig.android.feature.profile.aboutapp.AboutAppViewModel
 import com.hedvig.android.feature.profile.data.ProfileRepositoryDemo
 import com.hedvig.android.feature.profile.data.ProfileRepositoryImpl
@@ -70,4 +71,9 @@ val profileModule = module {
     MyInfoViewModel(get<ProfileRepositoryProvider>())
   }
   viewModel<AboutAppViewModel> { AboutAppViewModel(get<ApolloClient>()) }
+  single<CheckTravelCertificateDestinationAvailabilityUseCase> {
+    CheckTravelCertificateDestinationAvailabilityUseCaseImpl(
+      get<ApolloClient>(),
+    )
+  }
 }
