+++ app/feature/feature-profile/src/test/kotlin/com/hedvig/android/feature/profile/tab/ProfileViewModelTest.kt
@@ -12,8 +12,8 @@ import assertk.assertions.isNull
 import com.hedvig.android.auth.LogoutUseCase
 import com.hedvig.android.core.common.ErrorMessage
 import com.hedvig.android.core.common.test.MainCoroutineRule
-import com.hedvig.android.data.travelcertificate.CheckTravelCertificateDestinationAvailabilityUseCase
-import com.hedvig.android.data.travelcertificate.TravelCertificateAvailabilityError
+import com.hedvig.android.feature.profile.data.CheckTravelCertificateDestinationAvailabilityUseCase
+import com.hedvig.android.feature.profile.data.TravelCertificateAvailabilityError
 import com.hedvig.android.featureflags.flags.Feature
 import com.hedvig.android.featureflags.test.FakeFeatureManager
 import com.hedvig.android.featureflags.test.FakeFeatureManager2
@@ -381,7 +381,7 @@ private class FakeGetEurobonusStatusUseCase() : GetEurobonusStatusUseCase {
 }
 
 private class FakeCheckTravelCertificateDestinationAvailabilityUseCase :
-  CheckTravelCertificateDestinationAvailabilityUseCase {
+    CheckTravelCertificateDestinationAvailabilityUseCase {
   val turbine = Turbine<Either<TravelCertificateAvailabilityError, Unit>>(
     name = "TravelCertificateAvailability response",
   )
